# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: GigShield API
  description: Parametric Income Protection for Q-Commerce Delivery Partners
  version: 2.1.0
paths:
  /api/auth/login:
    post:
      tags:
      - Authentication
      summary: Login Worker
      description: Login an existing worker with phone and OTP.
      operationId: login_worker_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPVerify'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/register:
    post:
      tags:
      - Authentication
      summary: Register Worker
      description: Register a new delivery worker.
      operationId: register_worker_api_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerRegister'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/verify-otp:
    post:
      tags:
      - Authentication
      summary: Verify Otp
      description: 'Verify phone OTP. Mock: accepts ''1234'' for any phone.'
      operationId: verify_otp_api_auth_verify_otp_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPVerify'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/kyc:
    post:
      tags:
      - Authentication
      summary: Verify Kyc
      description: 'Aadhaar e-KYC verification. Mock: instant verification.'
      operationId: verify_kyc_api_auth_kyc_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KYCVerify'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/auth/setup-upi:
    post:
      tags:
      - Authentication
      summary: Setup Upi
      description: Setup UPI AutoPay mandate.
      operationId: setup_upi_api_auth_setup_upi_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UPISetup'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/auth/admin/login:
    post:
      tags:
      - Authentication
      summary: Admin Login
      description: Create an admin session for the operations portal.
      operationId: admin_login_api_auth_admin_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/me:
    get:
      tags:
      - Authentication
      summary: Get My Profile
      description: Get the authenticated worker profile.
      operationId: get_my_profile_api_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerResponse'
      security:
      - HTTPBearer: []
  /api/auth/profile/{worker_id}:
    get:
      tags:
      - Authentication
      summary: Get Profile
      description: Admin-only worker profile lookup.
      operationId: get_profile_api_auth_profile__worker_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: worker_id
        in: path
        required: true
        schema:
          type: integer
          title: Worker Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/workers:
    get:
      tags:
      - Authentication
      summary: List Workers
      description: List all registered workers.
      operationId: list_workers_api_auth_workers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
  /api/auth/account:
    delete:
      tags:
      - Authentication
      summary: Delete Account
      description: Delete/anonymise the authenticated worker account.
      operationId: delete_account_api_auth_account_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDeleteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
    AccountDeleteRequest:
      properties:
        otp:
          type: string
          title: Otp
        confirmation_text:
          type: string
          title: Confirmation Text
      type: object
      required:
      - otp
      - confirmation_text
      title: AccountDeleteRequest
    AdminLoginRequest:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
      - username
      - password
      title: AdminLoginRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    KYCVerify:
      properties:
        worker_id:
          type: integer
          title: Worker Id
        aadhaar_number:
          type: string
          pattern: ^\d{12}$
          title: Aadhaar Number
      type: object
      required:
      - worker_id
      - aadhaar_number
      title: KYCVerify
    OTPVerify:
      properties:
        phone:
          type: string
          title: Phone
        otp:
          type: string
          title: Otp
      type: object
      required:
      - phone
      - otp
      title: OTPVerify
...
    UPISetup:
      properties:
        worker_id:
          type: integer
          title: Worker Id
        upi_id:
          type: string
          title: Upi Id
      type: object
      required:
      - worker_id
      - upi_id
      title: UPISetup
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WorkerRegister:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 2
          title: Name
        phone:
          type: string
          pattern: ^[6-9]\d{9}$
          title: Phone
        platform:
          type: string
          title: Platform
          description: zepto, blinkit, dunzo, swiggy_instamart
        zone:
          type: string
          title: Zone
        city:
          type: string
          title: City
          default: Bengaluru
        vehicle_type:
          type: string
          title: Vehicle Type
          default: 2-wheeler
        partner_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Partner Id
        weekly_income:
          anyOf:
          - type: number
          - type: 'null'
          title: Weekly Income
          default: 5200.0
        accepted_terms:
          type: boolean
          title: Accepted Terms
          default: false
        accepted_privacy:
          type: boolean
          title: Accepted Privacy
          default: false
        accepted_automated_decisions:
          type: boolean
          title: Accepted Automated Decisions
          default: false
      type: object
      required:
      - name
      - phone
      - platform
      - zone
      title: WorkerRegister
    WorkerResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        phone:
          type: string
          title: Phone
        zone:
          type: string
          title: Zone
        city:
          type: string
          title: City
        platform:
          type: string
          title: Platform
        vehicle_type:
          type: string
          title: Vehicle Type
        upi_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Upi Id
        risk_tier:
          type: string
          title: Risk Tier
        weekly_income:
          type: number
          title: Weekly Income
        aadhaar_verified:
          type: boolean
          title: Aadhaar Verified
        phone_verified:
          type: boolean
          title: Phone Verified
        onboarding_complete:
          type: boolean
          title: Onboarding Complete
        is_deleted:
          type: boolean
          title: Is Deleted
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - phone
      - zone
      - city
      - platform
      - vehicle_type
      - upi_id
      - risk_tier
      - weekly_income
      - aadhaar_verified
      - phone_verified
      - onboarding_complete
      - is_deleted
      - created_at
      title: WorkerResponse
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer